requirements.txt

google-generativeai
python-dotenv
playwright
chromadb
